Fixed the frequency counter code.
I used floating point operation for the frequency counter. Could,
probably, just use a normal int an multiply by 10 or 100.  may need to
be changed, but it counts from 1hz to 2 kHz without problems.

+++ Project/EXTI.c
@@ -26,7 +26,7 @@ void InitEXTI()
 
 uint32_t time1IT=0; //wheel sensor 1 interrupt time/delay
 uint32_t time2IT=0; //wheel sensor 1 interrupt time/delay
-uint32_t Ws_deltat = 0; //delta time between trigger 1 and trigger 2
+float Ws_deltat = 0; //delta time between trigger 1 and trigger 2
 const uint8_t TRIGGER1=0; 
 const uint8_t TRIGGER2=1; 
 static uint8_t state = 0;
